own_requirements.txt
wandb
pytorch_wavelets
PyWavelets
cvxpy
timm
